requirements.txt
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
jsonschema>=4.17.0

# Database drivers (install as needed)
psycopg2-binary>=2.9.0  # PostgreSQL
mysql-connector-python>=8.0.0  # MySQL
# sqlite3 is included in Python standard library
